chore: test against latest lts only
This library is becoming less important with OAS3.1 being actual JSON Schema and with the frequency of Node releases I cant a) keep changing this matrix, or b) justify wasting compute resources on checking them all all the time. LMK if there's a problem in a particular version and we'll revert if it becomes an issue.